Update prefered ubuntu version to 22.04 (ships with python 3.10)
For latest versions of vantage6, python 3.10 or higher is required. Best
if it already comes from the official repositories (ubunutu), rather
than having to add external repos or resorting to other tricks to get a
newer version of python installed.

After dockerization story (v6) is done, we can drop this "python on the
host" requirement.

Ubuntu 22.04 is LTS (long time stable)