Update security.html
Remove aggregation node from list of required IPs to be allowlisted. This is because even in the case of having an aggregation node in a collaboration, all communication with it is done via the central server.

List harbor2.vantage6.ai, which will need to be used during install or updates of vantage6 (node software).